pyproject.toml
[project]
name = "ai-decision-council"
version = "1.2.0"
description = "Multi-model council orchestration with CLI and integration bridge"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "LLM Council Contributors"},
]
keywords = [
    "llm",
    "ai",
    "orchestration",
    "cli",
    "fastapi",
    "openrouter",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/pomoika/hackai26-pre-code"
Repository = "https://github.com/pomoika/hackai26-pre-code"
Issues = "https://github.com/pomoika/hackai26-pre-code/issues"

[project.scripts]
ai-decision-council = "ai_decision_council.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "pydantic>=2.9.0",
]
api = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_decision_council"]

[tool.hatch.build.targets.sdist]
include = [
    "src/ai_decision_council",
    "docs",
    "README.md",
    "CHANGELOG.md",
    "LICENSE",
    ".env.example",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--tb=short",
    "-q",
]

[tool.coverage.run]
source = ["src/ai_decision_council"]
omit = ["*/__pycache__/*", "*/py.typed"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
